[FIX] replace integer with float in validation
bzr revid: tpa@tinyerp.com-20131023113652-rj74krv1lecyp2xf

+++ addons/web_kanban_gauge/static/src/js/kanban_gauge.js
@@ -121,7 +121,7 @@ instance.web_kanban.GaugeWidget = instance.web_kanban.AbstractField.extend({
     },
 
     parse_client: function(value) {
-        return openerp.web.parse_value(value, { type:"integer" });
+        return openerp.web.parse_value(value, { type:"float" });
     },
 
 });
